default/requirements.txt
google-cloud-storage==1.13.2
tensorflow==1.12.0
numpy==1.15.4
scipy==1.2.0
pandas==0.23.4
sh